+++ pdf/test/annotations_test.dart
@@ -25,6 +25,7 @@ late Document pdf;
 void main() {
   setUpAll(() {
     Document.debug = true;
+    RichText.debug = true;
     pdf = Document();
   });