+++ pdf/test/colors_test.dart
@@ -173,6 +173,7 @@ late Document pdf;
 void main() {
   setUpAll(() {
     Document.debug = true;
+    RichText.debug = true;
     pdf = Document(pageMode: PdfPageMode.outlines);
   });