+++ pdf/test/orientation_test.dart
@@ -53,6 +53,7 @@ List<Widget> contentMultiPage(Context context) {
 void main() {
   setUpAll(() {
     Document.debug = true;
+    RichText.debug = true;
     pdf = Document();
   });