+++ pdf/test/widget_icon_test.dart
@@ -28,6 +28,7 @@ Font? icons;
 void main() {
   setUpAll(() {
     Document.debug = true;
+    RichText.debug = true;
     pdf = Document();
     icons = loadFont('material.ttf');
   });