+++ pdf/test/widget_table_test.dart
@@ -66,6 +66,7 @@ List<TableRow> buildTable(
 void main() {
   setUpAll(() {
     Document.debug = true;
+    RichText.debug = true;
     pdf = Document();
   });