+++ pdf/lib/src/widgets/decoration.dart
@@ -16,7 +16,6 @@
 
 import 'dart:math' as math;
 
-import 'package:image/image.dart' as im;
 import 'package:meta/meta.dart';
 import 'package:vector_math/vector_math_64.dart';
 
@@ -251,45 +250,6 @@ class BoxShadow {
   final PdfPoint offset;
   final double blurRadius;
   final double spreadRadius;
-
-  im.Image _rect(double width, double height) {
-    final shadow = im.Image(
-      (width + spreadRadius * 2).round(),
-      (height + spreadRadius * 2).round(),
-    );
-
-    im.fillRect(
-      shadow,
-      spreadRadius.round(),
-      spreadRadius.round(),
-      (spreadRadius + width).round(),
-      (spreadRadius + height).round(),
-      color.toInt(),
-    );
-
-    im.gaussianBlur(shadow, blurRadius.round());
-
-    return shadow;
-  }
-
-  im.Image _ellipse(double width, double height) {
-    final shadow = im.Image(
-      (width + spreadRadius * 2).round(),
-      (height + spreadRadius * 2).round(),
-    );
-
-    im.fillCircle(
-      shadow,
-      (spreadRadius + width / 2).round(),
-      (spreadRadius + height / 2).round(),
-      (width / 2).round(),
-      color.toInt(),
-    );
-
-    im.gaussianBlur(shadow, blurRadius.round());
-
-    return shadow;
-  }
 }
 
 enum BoxShape { circle, rectangle }
@@ -329,7 +289,8 @@ class BoxDecoration {
             if (borderRadius == null) {
               if (boxShadow != null) {
                 for (final s in boxShadow!) {
-                  final i = s._rect(box.width, box.height);
+                  final i = PdfRasterBase.shadowRect(box.width, box.height,
+                      s.spreadRadius, s.blurRadius, s.color);
                   final m = PdfImage.fromImage(context.document, image: i);
                   context.canvas.drawImage(
                     m,
@@ -342,7 +303,8 @@ class BoxDecoration {
             } else {
               if (boxShadow != null) {
                 for (final s in boxShadow!) {
-                  final i = s._rect(box.width, box.height);
+                  final i = PdfRasterBase.shadowRect(box.width, box.height,
+                      s.spreadRadius, s.blurRadius, s.color);
                   final m = PdfImage.fromImage(context.document, image: i);
                   context.canvas.drawImage(
                     m,
@@ -357,7 +319,8 @@ class BoxDecoration {
           case BoxShape.circle:
             if (boxShadow != null && box.width == box.height) {
               for (final s in boxShadow!) {
-                final i = s._ellipse(box.width, box.height);
+                final i = PdfRasterBase.shadowEllipse(box.width, box.height,
+                    s.spreadRadius, s.blurRadius, s.color);
                 final m = PdfImage.fromImage(context.document, image: i);
                 context.canvas.drawImage(
                   m,