+++ printing/lib/src/raster.dart
@@ -19,28 +19,16 @@ import 'dart:typed_data';
 import 'dart:ui' as ui;
 
 import 'package:flutter/painting.dart';
-import 'package:image/image.dart' as im;
+import 'package:pdf/pdf.dart';
 
 /// Represents a bitmap image
-class PdfRaster {
+class PdfRaster extends PdfRasterBase {
   /// Create a bitmap image
-  const PdfRaster(
-    this.width,
-    this.height,
-    this.pixels,
-  );
-
-  /// The width of the image
-  final int width;
-
-  /// The height of the image
-  final int height;
-
-  /// The raw RGBA pixels of the image
-  final Uint8List pixels;
-
-  @override
-  String toString() => 'Image ${width}x$height ${width * height * 4} bytes';
+  PdfRaster(
+    int width,
+    int height,
+    Uint8List pixels,
+  ) : super(width, height, true, pixels);
 
   /// Decode RGBA raw image to dart:ui Image
   Future<ui.Image> toImage() {
@@ -56,15 +44,11 @@ class PdfRaster {
   }
 
   /// Convert to a PNG image
+  @override
   Future<Uint8List> toPng() async {
     final image = await toImage();
     final data = await image.toByteData(format: ui.ImageByteFormat.png);
     return data!.buffer.asUint8List();
-  }
-
-  /// Returns the image as an [Image] object from the pub:image library
-  im.Image asImage() {
-    return im.Image.fromBytes(width, height, pixels);
   }
 }