+++ pdf/CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog
 
-## 3.8.5
+## 3.9.0
 
 - Improve TTF Writer compatibility
 - Apply THE BIDIRECTIONAL ALGORITHM using dart_bidi [Milad akarie]
 - Add Inseparable Widget
 - Fix unit tests
+- Update Image dependency
 
 ## 3.8.4